requirements.txt
pandas==2.3.3
clickhouse-connect
dbt-core
dbt-clickhouse
apache-airflow